Allow many more concurrent memcache connections
Previously we were limiting the number of memcache requests underway
because the client did not (appear to) support many concurrent
connections. But we are using an effectively dedicated memcached, and
memcached happily supports O(thousands) of connections and QPS.

Recent versions of the go client let you set the number of pooled
(i.e., persistent) connections to be higher than two, which we can
take advantage of.

It's sometimes not easy to see why the warmer is doing certain
work. To get a better idea, compile a list of missing things and a
list of imminently expiring things, and report on them individually,
before going and fetching them all.