Annotate resources caused by a FluxHelmRelease
We want to be able to identify the resources created, indirectly, by a
FluxHelmRelease.

We don't get access to the creation process, since that's managed by
Helm; but we do get to see, afterwards, which resources were involved,
in a big YAML string.

The simplest way to get the annotation done is to hand the big YAML
string directly to `kubectl annotate`.